Initial refactoring.
