+++ src/CMakeLists.txt
@@ -1,37 +1,35 @@
 # Main library
 
 set(library_sources
-  btwxt.h
-  btwxt.cpp
-  griddeddata.h
-  griddeddata.cpp
-  gridpoint.h
-  gridpoint.cpp
-  error.h
-)
+        regular-grid-interpolator-private.h
+        regular-grid-interpolator.cpp
+        grid-axis.cpp
+        )
 
-option( ${PROJECT_NAME}_STATIC_LIB "Make ${PROJECT_NAME} a static library" ON )
+option(${PROJECT_NAME}_STATIC_LIB "Make ${PROJECT_NAME} a static library" ON)
 
 if (${PROJECT_NAME}_STATIC_LIB)
-  add_library(${PROJECT_NAME} STATIC ${library_sources})
-  set_target_properties(${PROJECT_NAME} PROPERTIES COMPILE_FLAGS "-D${PROJECT_NAME}_STATIC_DEFINE")
-else()
-  set(CMAKE_MACOSX_RPATH 1)
-  add_library(${PROJECT_NAME} SHARED ${library_sources})
-endif()
+    add_library(${PROJECT_NAME} STATIC ${library_sources})
+    set_target_properties(${PROJECT_NAME} PROPERTIES COMPILE_FLAGS "-D${PROJECT_NAME}_STATIC_DEFINE")
+else ()
+    set(CMAKE_MACOSX_RPATH 1)
+    add_library(${PROJECT_NAME} SHARED ${library_sources})
+endif ()
 
 target_link_libraries(${PROJECT_NAME} PUBLIC courierr fmt)
 
+target_include_directories(${PROJECT_NAME} PUBLIC "${PROJECT_SOURCE_DIR}/include")
+
 target_compile_options(btwxt PRIVATE
-  $<$<CXX_COMPILER_ID:MSVC>:/W4>
-  $<$<OR:$<CXX_COMPILER_ID:Clang>,$<CXX_COMPILER_ID:AppleClang>,$<CXX_COMPILER_ID:GNU>>:
-  -Wall -Wextra -Wpedantic>
-)
+        $<$<CXX_COMPILER_ID:MSVC>:/W4>
+        $<$<OR:$<CXX_COMPILER_ID:Clang>,$<CXX_COMPILER_ID:AppleClang>,$<CXX_COMPILER_ID:GNU>>:
+        -Wall -Wextra -Wpedantic>
+        )
 
 target_compile_features(${PROJECT_NAME} PRIVATE cxx_std_17)
 include(GenerateExportHeader)
 generate_export_header(${PROJECT_NAME})
 
 if (${PROJECT_NAME}_COVERAGE)
-  add_coverage(${PROJECT_NAME})
-endif()
+    add_coverage(${PROJECT_NAME})
+endif ()
