+++ test/math_test.cpp
@@ -9,16 +9,16 @@
 #include "gtest/gtest.h"
 
 // btwxt
-#include "btwxt.h"
-#include "error.h"
-#include "griddeddata.h"
+#include <btwxt/btwxt.h>
+
+// testing
 #include "fixtures.hpp"
 
 using namespace Btwxt;
 
 // Unit tests on free functions, generally mathematically oriented.
 
-TEST(GriddedData, free_check_sorted) {
+TEST(GridAxis, free_check_sorted) {
   std::vector<std::pair<std::vector<double>, bool>> my_vecs = {{{1, 3, 5, 7, 9}, true},
                                                                {{1, 3, 5, 17, 9}, false},
                                                                {{9, 7, 5, 3, 1}, false},
@@ -36,11 +36,11 @@ TEST(GriddedData, locate_coords) {
   std::vector<std::vector<double>> grid = {{1, 2, 3, 4, 5}, {1, 2, 3, 4, 5, 6, 7}};
   std::vector<double> value1(5 * 7, 0.0);
   std::vector<std::vector<double>> values = {value1};
-  GriddedData test_gridded_data1(grid, values, courier);
+  RegularGridInterpolator test_rgi(grid, values, courier);
 
   std::vector<std::size_t> coords = {2, 3};
   std::vector<std::size_t> dimension_lengths = {5, 7};
-  std::size_t index = test_gridded_data1.get_value_index(coords);
+  std::size_t index = test_rgi.regular_grid_interpolator->get_value_index(coords);
   EXPECT_EQ(index, 17u);
 
   coords = {2, 3, 2};
@@ -48,7 +48,7 @@ TEST(GriddedData, locate_coords) {
 
   std::vector<double> value2(5 * 7 * 3, 0.0);
   values = {value2};
-  GriddedData test_gridded_data2(grid, values, courier);
+  RegularGridInterpolatorPrivate test_gridded_data2(grid, values, courier);
   index = test_gridded_data2.get_value_index(coords);
   EXPECT_EQ(index, 53u);
 }
@@ -63,49 +63,50 @@ TEST(GridPoint, set_dim_floor) {
 
   std::vector<double> target = {5.3};
 
-  GriddedData test_gridded_data = GriddedData({axis}, courier);
+  RegularGridInterpolator rgi({axis}, courier);
 
-  GridPoint grid_point(test_gridded_data, target, courier);
+  rgi.set_new_target(target);
 
-  grid_point.set_floor();
+  rgi.regular_grid_interpolator->set_floor();
 
-  EXPECT_EQ(grid_point.get_is_inbounds()[0], Bounds::INBOUNDS);
-  EXPECT_EQ(grid_point.get_floor()[0], 2u);
+  EXPECT_EQ(rgi.regular_grid_interpolator->get_is_inbounds()[0], Bounds::INBOUNDS);
+  EXPECT_EQ(rgi.regular_grid_interpolator->get_floor()[0], 2u);
   double expected_weight = 0.15;
-  double edge[2] = {axis.grid[grid_point.get_floor()[0]], axis.grid[grid_point.get_floor()[0] + 1]};
+  double edge[2] = {axis.grid[rgi.regular_grid_interpolator->get_floor()[0]],
+                    axis.grid[rgi.regular_grid_interpolator->get_floor()[0] + 1]};
   double returned_weight = compute_fraction(target[0], edge);
   EXPECT_DOUBLE_EQ(returned_weight, expected_weight);
 
-  grid_point.set_target({0.3});
-  grid_point.set_floor();
-  EXPECT_EQ(grid_point.get_is_inbounds()[0], Bounds::OUTBOUNDS);
-  EXPECT_EQ(grid_point.get_floor()[0], 0u);
+  rgi.regular_grid_interpolator->set_target({0.3});
+  rgi.regular_grid_interpolator->set_floor();
+  EXPECT_EQ(rgi.regular_grid_interpolator->get_is_inbounds()[0], Bounds::OUTBOUNDS);
+  EXPECT_EQ(rgi.regular_grid_interpolator->get_floor()[0], 0u);
 
-  grid_point.set_target({10.3});
-  grid_point.set_floor();
-  EXPECT_EQ(grid_point.get_is_inbounds()[0], Bounds::OUTBOUNDS);
-  EXPECT_EQ(grid_point.get_floor()[0], 3u);
+  rgi.regular_grid_interpolator->set_target({10.3});
+  rgi.regular_grid_interpolator->set_floor();
+  EXPECT_EQ(rgi.regular_grid_interpolator->get_is_inbounds()[0], Bounds::OUTBOUNDS);
+  EXPECT_EQ(rgi.regular_grid_interpolator->get_floor()[0], 3u);
 
-  grid_point.set_target({-0.3});
-  grid_point.set_floor();
-  EXPECT_EQ(grid_point.get_is_inbounds()[0], Bounds::OUTLAW);
-  EXPECT_EQ(grid_point.get_floor()[0], 0u);
+  rgi.regular_grid_interpolator->set_target({-0.3});
+  rgi.regular_grid_interpolator->set_floor();
+  EXPECT_EQ(rgi.regular_grid_interpolator->get_is_inbounds()[0], Bounds::OUTLAW);
+  EXPECT_EQ(rgi.regular_grid_interpolator->get_floor()[0], 0u);
 
-  grid_point.set_target({11.3});
-  grid_point.set_floor();
-  EXPECT_EQ(grid_point.get_is_inbounds()[0], Bounds::OUTLAW);
-  EXPECT_EQ(grid_point.get_floor()[0], 3u);
+  rgi.regular_grid_interpolator->set_target({11.3});
+  rgi.regular_grid_interpolator->set_floor();
+  EXPECT_EQ(rgi.regular_grid_interpolator->get_is_inbounds()[0], Bounds::OUTLAW);
+  EXPECT_EQ(rgi.regular_grid_interpolator->get_floor()[0], 3u);
 
-  test_gridded_data.grid_axes[0].set_extrap_limits({-DBL_MAX, DBL_MAX});
-  grid_point.set_target({-0.3});
-  grid_point.set_floor();
-  EXPECT_EQ(grid_point.get_is_inbounds()[0], Bounds::OUTBOUNDS);
-  EXPECT_EQ(grid_point.get_floor()[0], 0u);
+  rgi.set_axis_extrap_limits(0, {-DBL_MAX, DBL_MAX});
+  rgi.regular_grid_interpolator->set_target({-0.3});
+  rgi.regular_grid_interpolator->set_floor();
+  EXPECT_EQ(rgi.regular_grid_interpolator->get_is_inbounds()[0], Bounds::OUTBOUNDS);
+  EXPECT_EQ(rgi.regular_grid_interpolator->get_floor()[0], 0u);
 
-  grid_point.set_target({11.3});
-  grid_point.set_floor();
-  EXPECT_EQ(grid_point.get_is_inbounds()[0], Bounds::OUTBOUNDS);
-  EXPECT_EQ(grid_point.get_floor()[0], 3u);
+  rgi.regular_grid_interpolator->set_target({11.3});
+  rgi.regular_grid_interpolator->set_floor();
+  EXPECT_EQ(rgi.regular_grid_interpolator->get_is_inbounds()[0], Bounds::OUTBOUNDS);
+  EXPECT_EQ(rgi.regular_grid_interpolator->get_floor()[0], 3u);
 }
 
 TEST(GridPoint, compute_fraction) {
